Changed the approch that relies on folders in workspace to the epinio apps
Added an option to push application from source

Notifications improvements

General code cleanup, improvements

+++ src/extension.ts
@@ -1,10 +1,9 @@
 'use strict';
 import * as vscode from 'vscode';
-import { Application } from "./applications/models";
 
 import { WorkspaceConfigurator } from "./configurators/workspaceConfigurator";
 import { EpinioProvider } from "./explorers/providers";
-import { ApplicationNode, ApplicationsNode } from "./applications/views";
+import { ApplicationNode } from "./applications/views";
 import { ServiceNode } from "./services/views";
 
 export function activate(context: vscode.ExtensionContext) {
@@ -26,46 +25,86 @@ export function activate(context: vscode.ExtensionContext) {
     vscode.commands.registerCommand('epinio.showOutput', () => outputChannel.show());
     disposables.push(outputChannel);
 
-    let refreshExplorer = vscode.commands.registerCommand("epinio.application.explorer.refresh", () => {
+    const addExplorer = vscode.commands.registerCommand("epinio.application.explorer.add", () => {
+        vscode.window.showOpenDialog({
+            title: 'Epinio: Select Application source folder',
+            canSelectFiles: true,
+            canSelectFolders: true
+        }).then(res => {
+            if(res) {
+                provider.pushApplicationFromSource(res[0]?.path.split('/').pop() || 'default-name', res[0]?.fsPath);
+            }
+        });
+    });
+
+    const refreshExplorer = vscode.commands.registerCommand("epinio.application.explorer.refresh", () => {
         provider.refresh();
     });
 
-    let pushApplication = vscode.commands.registerCommand("epinio.application.push", (node: ApplicationNode) => {
-        provider.pushApplication(node);
+    const pushApplication = vscode.commands.registerCommand("epinio.application.push", (node: ApplicationNode) => {
+        vscode.window.showOpenDialog({
+            title: 'Epinio: Select Application source folder',
+            canSelectFiles: true,
+            canSelectFolders: true
+        }).then(res => {
+            if(res) {
+                provider.pushApplication(node, res[0]?.fsPath);
+            }
+        });
     });
 
-    let openApplication = vscode.commands.registerCommand("epinio.application.open", (node: ApplicationNode) => {
+    const openApplication = vscode.commands.registerCommand("epinio.application.open", (node: ApplicationNode) => {
         provider.openApplication(node);
     });
 
-    let applicationEnv = vscode.commands.registerCommand("epinio.application.env", (node: ApplicationNode) => {
-        provider.applicationEnv(node);
+    const applicationEnv = vscode.commands.registerCommand("epinio.application.env", (node: ApplicationNode) => {
+        provider.setApplicationEnv(node);
     });
 
-    let scaleApplication = vscode.commands.registerCommand("epinio.application.scale", (node: ApplicationNode) => {
+    const scaleApplication = vscode.commands.registerCommand("epinio.application.scale", (node: ApplicationNode) => {
         provider.scaleApplication(node);
     });
 
-    let applicationLogs = vscode.commands.registerCommand("epinio.application.logs", (node: ApplicationNode) => {
+    const applicationLogs = vscode.commands.registerCommand("epinio.application.logs", (node: ApplicationNode) => {
         provider.applicationLogs(node);
     });
 
-    let deleteApplication = vscode.commands.registerCommand("epinio.application.delete", (node: ApplicationNode) => {
-        provider.deleteApplication(node);
+    const deleteApplication = vscode.commands.registerCommand("epinio.application.delete",  (node: ApplicationNode) => {
+        vscode.window.showInformationMessage(
+            `Are you sure you want to delete the app?`,
+            { modal: true },
+            ...['Yes', 'No'],
+        ).then(res =>  {
+            if(res === 'Yes') {
+                provider.deleteApplication(node);
+/*                 vscode.window.withProgress({
+                    location: vscode.ProgressLocation.Notification,
+                    cancellable: false
+                }, async (progress, token) => {     
+                    progress.report({
+                        message: `Deleting Application...`,
+                    });               
+                    provider.deleteApplication(node).then(res => {
+                        vscode.window.showInformationMessage('Application deleted successfully!');
+                    })
+                }); */
+            } 
+        });
     });
 
-    let bindService = vscode.commands.registerCommand("epinio.service.bind", (node: ServiceNode) => {
+    const bindService = vscode.commands.registerCommand("epinio.service.bind", (node: ServiceNode) => {
         provider.bindService(node);
     });
 
-    let unbindService = vscode.commands.registerCommand("epinio.service.unbind", (node: ServiceNode) => {
+    const unbindService = vscode.commands.registerCommand("epinio.service.unbind", (node: ServiceNode) => {
         provider.unbindService(node);
     });
 
-    let deleteService = vscode.commands.registerCommand("epinio.service.delete", (node: ServiceNode) => {
+    const deleteService = vscode.commands.registerCommand("epinio.service.delete", (node: ServiceNode) => {
         provider.deleteService(node);
     });
 
+    context.subscriptions.push(addExplorer);
     context.subscriptions.push(refreshExplorer);
     context.subscriptions.push(pushApplication);
     context.subscriptions.push(openApplication);
